Split GitHub contribution responsibilities
